requirements.txt
mysql-connector-python
python-dotenv
fastapi
uvicorn
starlette
dotenv
pydantic
python-jose